Handle returning promises with async.
The waiter was not properly handling `fn` returning a promise that had
async. This updates to ensure that any returned promise is resolved
before marking the waiter as completed.

---

The existing tests were using async (via `setTimeout`) to kick off the
callback, but since the promise used was resolving synchronously in the
`Ember.RSVP.Promise` constructor true async from the promise itself was
not being tested (but the async from the `setTimeout` delaying the
promise creation was).