Ensure callback is invoked immediately.
Prior to this change we would add a test waiter that would not complete
until the wrapped callback was invoked, but we would never actually
invoke it.

The majority of other `run.*` callback methods (with the exception of
`Ember.run.bind`) will always ensure that the callback provided is
invoked (i.e. `Ember.run`, `Ember.run.next`, `Ember.run.later`, etc).

This changes to ensure that the callback is invoked.

---

I could easily see where we might want to ensure that we always invoke
in the next tick (to ensure an async paradigm), but I'm not certain...